flask-microservice/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0

# Environment & Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2

# Machine Learning
scikit-learn==1.3.1
tensorflow==2.13.0
mediapipe==0.10.11

# Computer Vision
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
Pillow==10.0.0

# Data Validation - DOWNGRADED to v1.x for TensorFlow 2.13.0 compatibility
pydantic==1.10.13

# Critical: Pin typing-extensions to satisfy both TensorFlow and Pydantic 1.x
typing-extensions==4.5.0